+++ packages/themes/src/pieces/fontSystemUI.ts
@@ -1,0 +1,9 @@
+import { ThemeSpec } from '@chsk/core'
+
+export const fontSystemUIThemePiece: ThemeSpec = {
+    text: {
+        default: {
+            fontFamily: 'system-ui',
+        },
+    },
+}