+++ packages/themes/src/pieces/index.ts
@@ -1,1 +1,4 @@
 export * from './download'
+export * from './faintTicks'
+export * from './fontSystemUI'
+export * from './smallAxisText'