+++ packages/themes/tests/complete/inverseGridTheme.test.tsx
@@ -1,4 +1,4 @@
-import { inverseGridTheme } from '../src'
+import { inverseGridTheme } from '../../src'
 
 describe('inverse Grid theme', () => {
     it('defines styles for line and rect', () => {